+++ src/decorators/password-blacklist.decorator.ts
@@ -1,10 +1,10 @@
 import { registerDecorator } from "class-validator";
-import { PasswordConstraint } from "./password.constraint";
-import { PasswordValidatorOptions } from "./password-validator-options.interface";
+import { OptionsBlacklistRule } from "../password-validator-options.interface";
 import { ValidationOptions } from "class-validator/types/decorator/ValidationOptions";
+import { PasswordBlacklistConstraint } from "../constraints/password-blacklist.constraint";
 
-export function IsPassword(
-    ValidatorOptions?: PasswordValidatorOptions,
+export function PasswordBlacklist(
+    validatorOptions?: OptionsBlacklistRule,
     defaultValidatorOptions?: ValidationOptions
 ) {
     return function (object: object, propertyName: string) {
@@ -12,8 +12,8 @@ export function IsPassword(
             target: object.constructor,
             propertyName: propertyName,
             options: defaultValidatorOptions,
-            constraints: [ValidatorOptions],
-            validator: PasswordConstraint,
+            constraints: [validatorOptions],
+            validator: PasswordBlacklistConstraint,
         });
     };
 }