+++ src/password-validator-options.interface.ts
@@ -31,5 +31,16 @@ export interface HasRule {
 export interface ConfirmPasswordRule {
     message?: string;
     enabled: boolean;
-    compareToProperty?: string;
+    compareToProperty: string;
+}
+export type OptionsLengthRule = Omit<LengthRule, "enabled">;
+export interface OptionsConfirmPasswordRule
+    extends Omit<ConfirmPasswordRule, "enabled"> {
+    compareToProperty: string;
+}
+export type OptionsBlacklistRule = Omit<BlacklistRule, "enabled">;
+
+export interface OptionsHasSpecialCharsRule
+    extends Omit<HasSpecialCharsRule, "enabled"> {
+    characters: string[];
 }